+++ src/anomalib/data/datamodules/base/__init__.py
@@ -1,17 +1,9 @@
-"""Base classes for custom dataset and datamodules."""
+"""Base Anomalib Data Modules."""
 
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .datamodule import AnomalibDataModule
-from .dataset import AnomalibDataset
-from .depth import AnomalibDepthDataset
-from .video import AnomalibVideoDataModule, AnomalibVideoDataset
+from .image import AnomalibDataModule
+from .video import AnomalibVideoDataModule
 
-__all__ = [
-    "AnomalibDataset",
-    "AnomalibDataModule",
-    "AnomalibVideoDataset",
-    "AnomalibVideoDataModule",
-    "AnomalibDepthDataset",
-]
+__all__ = ["AnomalibDataModule", "AnomalibVideoDataModule"]
