+++ src/anomalib/data/datamodules/depth/__init__.py
@@ -1,4 +1,4 @@
-"""Anomalib Depth Datasets."""
+"""Anomalib Depth Data Modules."""
 
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
