+++ src/anomalib/data/datamodules/depth/mvtec_3d.py
@@ -1,4 +1,4 @@
-"""MVTec 3D-AD Dataset (CC BY-NC-SA 4.0).
+"""MVTec 3D-AD Datamodule (CC BY-NC-SA 4.0).
 
 Description:
     This script contains PyTorch Dataset, Dataloader and PyTorch Lightning DataModule for the MVTec 3D-AD dataset.
@@ -16,33 +16,27 @@ Reference:
         0010865000003124.
 """
 
-# Copyright (C) 2022 Intel Corporation
+# Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 import logging
-from collections.abc import Sequence
 from pathlib import Path
 
-from pandas import DataFrame
 from torchvision.transforms.v2 import Transform
 
 from anomalib import TaskType
-from anomalib.data.datamodules.base import AnomalibDataModule, AnomalibDepthDataset
-from anomalib.data.errors import MisMatchError
+from anomalib.data.datamodules.base import AnomalibDataModule
+from anomalib.data.datasets import MVTec3DDataset
 from anomalib.data.utils import (
     DownloadInfo,
-    LabelName,
     Split,
     TestSplitMode,
     ValSplitMode,
     download_and_extract,
-    validate_path,
 )
 
 logger = logging.getLogger(__name__)
 
-
-IMG_EXTENSIONS = [".png", ".PNG", ".tiff"]
 
 DOWNLOAD_INFO = DownloadInfo(
     name="mvtec_3d",
@@ -50,158 +44,6 @@ DOWNLOAD_INFO = DownloadInfo(
     "/mvtec_3d_anomaly_detection.tar.xz",
     hashsum="d8bb2800fbf3ac88e798da6ae10dc819",
 )
-
-CATEGORIES = ("bagel", "cable_gland", "carrot", "cookie", "dowel", "foam", "peach", "potato", "rope", "tire")
-
-
-def make_mvtec_3d_dataset(
-    root: str | Path,
-    split: str | Split | None = None,
-    extensions: Sequence[str] | None = None,
-) -> DataFrame:
-    """Create MVTec 3D-AD samples by parsing the MVTec AD data file structure.
-
-    The files are expected to follow this structure:
-    - `path/to/dataset/split/category/image_filename.png`
-    - `path/to/dataset/ground_truth/category/mask_filename.png`
-
-    This function creates a DataFrame to store the parsed information. The DataFrame follows this format:
-
-    +---+---------------+-------+---------+---------------+---------------------------------------+-------------+
-    |   | path          | split | label   | image_path    | mask_path                             | label_index |
-    +---+---------------+-------+---------+---------------+---------------------------------------+-------------+
-    | 0 | datasets/name | test  | defect  | filename.png  | ground_truth/defect/filename_mask.png | 1           |
-    +---+---------------+-------+---------+---------------+---------------------------------------+-------------+
-
-    Args:
-        root (Path): Path to the dataset.
-        split (str | Split | None, optional): Dataset split (e.g., 'train' or 'test').
-            Defaults to ``None``.
-        extensions (Sequence[str] | None, optional): List of file extensions to be included in the dataset.
-            Defaults to ``None``.
-
-    Examples:
-        The following example shows how to get training samples from the MVTec 3D-AD 'bagel' category:
-
-        >>> from pathlib import Path
-        >>> root = Path('./MVTec3D')
-        >>> category = 'bagel'
-        >>> path = root / category
-        >>> print(path)
-        PosixPath('MVTec3D/bagel')
-
-        >>> samples = create_mvtec_3d_ad_samples(path, split='train')
-        >>> print(samples.head())
-            path          split label image_path                          mask_path                        label_index
-            MVTec3D/bagel train good MVTec3D/bagel/train/good/rgb/105.png MVTec3D/bagel/ground_truth/good/gt/105.png 0
-            MVTec3D/bagel train good MVTec3D/bagel/train/good/rgb/017.png MVTec3D/bagel/ground_truth/good/gt/017.png 0
-
-    Returns:
-        DataFrame: An output DataFrame containing the samples of the dataset.
-    """
-    if extensions is None:
-        extensions = IMG_EXTENSIONS
-
-    root = validate_path(root)
-    samples_list = [(str(root),) + f.parts[-4:] for f in root.glob(r"**/*") if f.suffix in extensions]
-    if not samples_list:
-        msg = f"Found 0 images in {root}"
-        raise RuntimeError(msg)
-
-    samples = DataFrame(samples_list, columns=["path", "split", "label", "type", "file_name"])
-
-    # Modify image_path column by converting to absolute path
-    samples.loc[(samples.type == "rgb"), "image_path"] = (
-        samples.path + "/" + samples.split + "/" + samples.label + "/" + "rgb/" + samples.file_name
-    )
-    samples.loc[(samples.type == "rgb"), "depth_path"] = (
-        samples.path
-        + "/"
-        + samples.split
-        + "/"
-        + samples.label
-        + "/"
-        + "xyz/"
-        + samples.file_name.str.split(".").str[0]
-        + ".tiff"
-    )
-
-    # Create label index for normal (0) and anomalous (1) images.
-    samples.loc[(samples.label == "good"), "label_index"] = LabelName.NORMAL
-    samples.loc[(samples.label != "good"), "label_index"] = LabelName.ABNORMAL
-    samples.label_index = samples.label_index.astype(int)
-
-    # separate masks from samples
-    mask_samples = samples.loc[((samples.split == "test") & (samples.type == "rgb"))].sort_values(
-        by="image_path",
-        ignore_index=True,
-    )
-    samples = samples.sort_values(by="image_path", ignore_index=True)
-
-    # assign mask paths to all test images
-    samples.loc[((samples.split == "test") & (samples.type == "rgb")), "mask_path"] = (
-        mask_samples.path + "/" + samples.split + "/" + samples.label + "/" + "gt/" + samples.file_name
-    )
-    samples = samples.dropna(subset=["image_path"])
-    samples = samples.astype({"image_path": "str", "mask_path": "str", "depth_path": "str"})
-
-    # assert that the right mask files are associated with the right test images
-    mismatch_masks = (
-        samples.loc[samples.label_index == LabelName.ABNORMAL]
-        .apply(lambda x: Path(x.image_path).stem in Path(x.mask_path).stem, axis=1)
-        .all()
-    )
-    if not mismatch_masks:
-        msg = """Mismatch between anomalous images and ground truth masks. Make sure the mask files
-          in 'ground_truth' folder follow the same naming convention as the anomalous images in
-          the dataset (e.g. image: '000.png', mask: '000.png' or '000_mask.png')."""
-        raise MisMatchError(msg)
-
-    mismatch_depth = (
-        samples.loc[samples.label_index == LabelName.ABNORMAL]
-        .apply(lambda x: Path(x.image_path).stem in Path(x.depth_path).stem, axis=1)
-        .all()
-    )
-    if not mismatch_depth:
-        msg = """Mismatch between anomalous images and depth images. Make sure the mask files in
-          'xyz' folder follow the same naming convention as the anomalous images in the dataset
-          (e.g. image: '000.png', depth: '000.tiff')."""
-        raise MisMatchError(msg)
-
-    if split:
-        samples = samples[samples.split == split].reset_index(drop=True)
-
-    return samples
-
-
-class MVTec3DDataset(AnomalibDepthDataset):
-    """MVTec 3D dataset class.
-
-    Args:
-        task (TaskType): Task type, ``classification``, ``detection`` or ``segmentation``
-        root (Path | str): Path to the root of the dataset
-            Defaults to ``"./datasets/MVTec3D"``.
-        category (str): Sub-category of the dataset, e.g. 'bagel'
-            Defaults to ``"bagel"``.
-        transform (Transform, optional): Transforms that should be applied to the input images.
-            Defaults to ``None``.
-        split (str | Split | None): Split of the dataset, usually Split.TRAIN or Split.TEST
-            Defaults to ``None``.
-    """
-
-    def __init__(
-        self,
-        task: TaskType,
-        root: Path | str = "./datasets/MVTec3D",
-        category: str = "bagel",
-        transform: Transform | None = None,
-        split: str | Split | None = None,
-    ) -> None:
-        super().__init__(task=task, transform=transform)
-
-        self.root_category = Path(root) / Path(category)
-        self.split = split
-        self.samples = make_mvtec_3d_dataset(self.root_category, split=self.split, extensions=IMG_EXTENSIONS)
 
 
 class MVTec3D(AnomalibDataModule):
