+++ src/anomalib/data/datamodules/image/__init__.py
@@ -1,7 +1,4 @@
-"""Anomalib Image Datasets.
-
-This module contains the supported image datasets for Anomalib.
-"""
+"""Anomalib Image Data Modules."""
 
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
