+++ tests/unit/data/utils/test_synthetic.py
@@ -9,7 +9,7 @@ from pathlib import Path
 import pytest
 
 from anomalib import TaskType
-from anomalib.data.datamodules.image.folder import FolderDataset
+from anomalib.data.datasets.image.folder import FolderDataset
 from anomalib.data.utils.synthetic import SyntheticAnomalyDataset
